fix(desktop): isolate project cache state from gateway mocks
profile-projects-boundary imported projects.ts, which pulls activeGateway
and broke unit tests that only partially mock @/store/gateway.

Move list/tree/scope atoms and clearProjectsCacheForProfileSwitch into
projects-cache-state (no gateway imports) so profile load stays lightweight.
