+++ apps/desktop/src/store/profile-projects-boundary.ts
@@ -1,21 +1,8 @@
 /**
  * Boundary helpers between profile switching and the projects cache.
  *
- * Kept in its own module so `profile.ts` can clear project state without a
- * circular import through `projects.ts` (which already imports profile).
+ * Imports only gateway-free project state so unit tests that mock part of
+ * `@/store/gateway` can load `profile.ts` without needing a full gateway mock.
  */
 
-import {
-  $activeProjectId,
-  $projects,
-  $projectTree,
-  exitProjectScope
-} from '@/store/projects'
-
-/** Drop the previous profile's projects.db snapshot on profile switch (#79406). */
-export function clearProjectsCacheForProfileSwitch(): void {
-  $projects.set([])
-  $projectTree.set([])
-  $activeProjectId.set(null)
-  exitProjectScope()
-}
+export { clearProjectsCacheForProfileSwitch } from '@/store/projects-cache-state'
